model-bench/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "model-bench"
version = "0.1.0"
description = "Worker model benchmarking tool — run tasks across models, score with Opus, compare."
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.50.0",
    "httpx>=0.27.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
model_bench = "model_bench.cli:app"
